fix(scope): Prevent duplicate acquisitions
When the period between laser pulses is longer than the period between
measurements, it's possible to accidentally transfer data that has
already been transferred. This set of changes makes it such that the
experiment worker will not transfer data unless it sees that the data is
from a new acquisition. It also eliminates the need to transfer the
shutter channel, but leaves that for future work.